Split the one binary to two
